fix: renamed module ioutil -> httputil
